CaseStudyCreator/pyproject.toml
[project]
name = "RecommendationEngine"
version = "0.1.0"
description = "RecommendationEngine using crewAI"
authors = [{ name = "Vladimir", email = "vladimir@elementsagents.com" }]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "crewai-tools>=0.32.1",
    "crewai[tools]>=0.86.0,<1.0.0",
    "docling>=2.15.1",
    "weaviate-client>=4.10.4",
    "youtube-transcript-api>=0.6.2",
    "aiosqlite>=0.19.0",
    "embedchain>=0.1.124"
]

[project.scripts]
CaseStudyCreator = "CaseStudyCreator.main:run"
run_crew = "CaseStudyCreator.main:run"
train = "CaseStudyCreator.main:train"
replay = "CaseStudyCreator.main:replay"
test = "CaseStudyCreator.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build]
include = [
    "src/**/*.py",
    "src/**/*.yaml",
    "pyproject.toml",
    "README.md"
]